Moved lowdb stuff into new service class
